Netlist: - Working Ebers Moll model. That's a significant step ahead. - Simple 2x2|RHS SPICE stamps now supported by two terminal devices.   This was implicitly contained in the design, but set_mat now shows how   a two-terminal device fits into a SPICE approach. - Introduced direct solvers for net groups with 1 or 2 nets. - Introduced specialized solvers for N=1,2,3,4,5 by using templates. - nice performance increase for BJTs

+++ src/emu/netlist/nl_setup.c
@@ -188,10 +188,10 @@ void netlist_setup_t::register_object(netlist_device_t &dev, const pstring &name
 						case netlist_param_t::LOGIC:
 						{
 							NL_VERBOSE_OUT(("Found parameter ... %s : %s\n", name.cstr(), val->cstr()));
-							int vald = 0;
-							if (sscanf(val.cstr(), "%d", &vald) != 1)
+							double vald = 0;
+							if (sscanf(val.cstr(), "%lf", &vald) != 1)
 								netlist().error("Invalid number conversion %s : %s\n", name.cstr(), val.cstr());
-							dynamic_cast<netlist_param_int_t &>(param).initial(vald);
+							dynamic_cast<netlist_param_int_t &>(param).initial((int) vald);
 						}
 						break;
 						case netlist_param_t::STRING:
